tictactoe
Tic-tac-toe is a game in which two players take turns in drawing either an ' O' or an ' X' in one square of a grid consisting of nine squares. The winner is the first player to get three of the same symbols in a row. This is a not GUI representation of the game hat can be played using command prompt